refactor(Routes & Action creators):  Work under construction 🚧
- Public post action creators organized
