+++ client/src/components/PostCreateForm/PostCreateForm.tsx
@@ -1,12 +1,8 @@
-import { FC, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { Navigate, useNavigate, useParams } from "react-router";
-import {
-  createPost,
-  getPosts,
-  getPostsById,
-} from "../../redux/actions/actionCreators";
-import { editPost } from "../../redux/actions/companyActionCreators";
+import {FC, useState} from "react";
+import {useDispatch, useSelector} from "react-redux";
+import { useNavigate, useParams} from "react-router";
+import { createPost } from "../../redux/actions/actionCreators";
+import {editPost} from "../../redux/actions/companyActionCreators";
 import PostCreateModal from "../PostCreateModal";
 import {
   Container,
@@ -45,10 +41,10 @@ type Props = {
   mode: string;
 };
 
-const PostCreateForm: FC<Props> = ({ mode }) => {
+const PostCreateForm: FC<Props> = ({mode}) => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
-  const { postId, companyId } = useParams();
+  const {postId, companyId} = useParams();
 
   const postCreateModal = useSelector(
     (state: any) => state.postsReducer.postCreateModal
@@ -78,19 +74,19 @@ const PostCreateForm: FC<Props> = ({ mode }) => {
   const [tag, setTag] = useState<Tag>("");
 
   const addTag = () => {
-    setForm({ ...form, tags: [...form.tags, tag] });
+    setForm({...form, tags: [...form.tags, tag]});
     setTag("");
   };
 
-  const deleteTag = ({ target: { name } }: any) => {
-    setForm({ ...form, tags: form.tags.filter((e) => e !== name) });
+  const deleteTag = ({target: {name}}: any) => {
+    setForm({...form, tags: form.tags.filter((e) => e !== name)});
   };
 
-  const handleInputs = ({ target: { name, value } }: any) => {
+  const handleInputs = ({target: {name, value}}: any) => {
     if (name === "tag") {
       setTag(value);
     } else {
-      setForm({ ...form, [name]: value });
+      setForm({...form, [name]: value});
     }
   };
