+++ client/src/components/PostDetail.tsx
@@ -2,7 +2,7 @@ import { useEffect, FC } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useNavigate, useParams } from "react-router";
 import styled from "styled-components";
-import { getPostsById } from "../redux/actions/actionCreators";
+import { getPostsById } from "../redux/actions/public/postsActions";
 import CompanyLogo from "../assets/company-logo.svg";
 import { Link } from "react-router-dom";
 import { jobApplication } from "../redux/actions/applicantActionCreators";
